+++ src/Metadata/PropertyTypeIterable.php
@@ -6,7 +6,6 @@ namespace Liip\MetadataParser\Metadata;
 
 use Symfony\Component\TypeInfo\Type\CollectionType;
 use Symfony\Component\TypeInfo\Type\GenericType;
-use Symfony\Component\TypeInfo\Type\ObjectType;
 
 /**
  * This property type can be merged with PropertyTypeClass<T>, provided that T is, inherits from, or is a parent class of {@see PropertyTypeIterable::traversableClass}
@@ -36,13 +35,16 @@ final class PropertyTypeIterable extends AbstractPropertyType
 
     public function isHashmap(): bool
     {
-        $wrapped = $this->typeInfo->getWrappedType();
-
-        if ($wrapped instanceof GenericType && $wrapped->getWrappedType() instanceof ObjectType) {
-            return 2 === \count($wrapped->getVariableTypes());
+        if ($this->typeInfo->isList()) {
+            return false;
         }
 
-        return !$this->typeInfo->isList();
+        $wrapped = $this->typeInfo->getWrappedType();
+        if (!$wrapped instanceof GenericType) {
+            return false;
+        }
+
+        return \count($wrapped->getVariableTypes()) > 1;
     }
 
     /**